% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_dist.R
\name{comp_dist}
\alias{comp_dist}
\title{Squared euclidean distance for matrices}
\usage{
comp_dist(x1, x2)
}
\arguments{
\item{x1}{Matrix}

\item{x2}{Matrix}
}
\value{
Squared euclidean distance between points in x1 and x2
}
\description{
Squared euclidean distance for matrices
}
\examples{
x1 <- matrix(rnorm(20), 10, 2)
x2 <- matrix(rnorm(20), 10, 2)
}
