% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_median.R
\name{comp_med}
\alias{comp_med}
\title{Compute median squared euclidean distance}
\usage{
comp_med(x)
}
\arguments{
\item{x}{Matrix}
}
\value{
Scalar: median squared euclidean distance
}
\description{
Compute median squared euclidean distance
}
\examples{
x <- matrix(sample(1:20), 5, 4)
comp_med(x)
}
