% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_median.R
\name{comp_median}
\alias{comp_median}
\title{Compute median squared euclidean distance}
\usage{
comp_median(x)
}
\arguments{
\item{x}{Matrix}
}
\value{
Scalar: median squared euclidean distance
}
\description{
Compute median squared euclidean distance
}
