% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian_kernel.R
\name{gaussian_kernel}
\alias{gaussian_kernel}
\title{Gaussian kernel function}
\usage{
gaussian_kernel(distance_matrix, sigma)
}
\arguments{
\item{distance_matrix}{Matrix of squared euclidean matrices}

\item{sigma}{Kernel bandwidth parameter}
}
\value{
Gaussian kernel applied to the input matrix
}
\description{
Gaussian kernel function
}
