% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_lambda_grid_search.R
\name{sigma_lambda_grid_search}
\alias{sigma_lambda_grid_search}
\title{Grid search via cross validation on sigma and lambda parameters}
\usage{
sigma_lambda_grid_search(
  Nnu,
  Nde,
  k,
  dist_nu,
  dist_de,
  sigma_list,
  lambda_list,
  alpha,
  n_folds
)
}
\arguments{
\item{Nnu}{Number of points in numerator}

\item{Nde}{Number of points in denomenator}

\item{k}{Number of basis functions}

\item{dist_nu}{Distance matrix for numerator}

\item{dist_de}{Distance matrix for denomenator}

\item{sigma_list}{Sigma values to search over}

\item{lambda_list}{Lambda values to search over}

\item{alpha}{Relative parameter}

\item{n_folds}{Number of folds for CV}
}
\value{
Optimal sigma-lambda pair
}
\description{
Grid search via cross validation on sigma and lambda parameters
}
