% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{sliding_window}
\alias{sliding_window}
\title{Matrix of subsequences (sliding window)}
\usage{
sliding_window(X, window_size)
}
\arguments{
\item{X}{Data matrix}

\item{window_size}{How many subsequences constitute a single window}
}
\value{
Matrix of subsequences
}
\description{
Matrix of subsequences (sliding window)
}
