% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_detect.R
\name{ts_detect}
\alias{ts_detect}
\title{Change point detection}
\usage{
ts_detect(ts, window_size = 5, step = NULL, alpha = 0.05, k = 100, n_folds = 5)
}
\arguments{
\item{ts}{Time series to detect change points in. Assumes this is a D by N matrix
where D is the dimension of the time series and N is the number of
time points.}

\item{window_size}{The length of the sub-sequences generated from the series. Default \code{5}.}

\item{step}{How many sub-sequences forward and backward to from a time point
to compute a score from. Default is 10\% of the length of the series
if not specified.}

\item{alpha}{Relative parameter. Default \code{0.05}. Setting to \code{0} recovers
ordinary unconstrained least squares importance fitting.}

\item{k}{Number of basis functions. Default is minimum of \code{100} and dimension of
the time series.}

\item{n_folds}{Number of folds to use in determining optimal kernel bandwidth
and lambda parameter in RULSIF.}
}
\value{
A vector of change point scores for each time point starting at t = step
and ending at t = (number of time points) - window_size.
}
\description{
Detect change points in a time series using RelULSIF.
}
\examples{
t <- c(rnorm(150, mean = 0), rnorm(150, mean = 5), rnorm(150, mean = 1))
ts_detect(t)
}
