% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{ts_plot}
\alias{ts_plot}
\title{Plot change point detection results}
\usage{
ts_plot(ts, step, scores, change_points)
}
\arguments{
\item{ts}{Time series input}

\item{step}{Starting point of detection}

\item{scores}{rPE scores}

\item{change_points}{Detected change point indices}
}
\description{
Plot change point detection results
}
