% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RelULSIF.R
\name{RelULSIF}
\alias{RelULSIF}
\title{Relative unconstrained least squares importance fitting}
\usage{
RelULSIF(
  Xnu,
  Xde,
  Xce = NULL,
  sigma = NULL,
  lambda = NULL,
  alpha = 0.01,
  k = 100,
  n_folds = 5
)
}
\arguments{
\item{Xnu}{Samples from numerator probability density}

\item{Xde}{Samples from denomenator probability density}

\item{Xce}{Matrix of centers}

\item{sigma}{Scalar or vector; Gaussian kernel bandwidth(s)}

\item{lambda}{Scalar or vector; regularization parameter(s)}

\item{alpha}{Scalar; relative parameter in [0, 1)}

\item{k}{Integer; number of basis functions}

\item{n_folds}{Integer; number of folds to use in cross fold validation}
}
\value{
List of 3
\itemize{
\item opt_sigma: chosen sigma from CV
\item opt_lambda: chosen lambda from CV
\item rPE: relative Pearson divergence
}
}
\description{
Relative unconstrained least squares importance fitting
}
